LPS-199789 Move keyboard movement components to new folder
