+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/keyboard_movement/KeyboardMovementText.js
@@ -8,7 +8,7 @@ import React, {useEffect} from 'react';
 import {
 	useMovementText,
 	useSetMovementText,
-} from '../contexts/KeyboardMovementContext';
+} from '../../contexts/KeyboardMovementContext';
 
 export default function KeyboardMovementText() {
 	const text = useMovementText();
